Rename github folder
